LRCI-5369 Consolidate logic
