+++ modules/test/jenkins-results-parser/src/test/java/com/liferay/jenkins/results/parser/test/suite/BaseRelevantRuleTestCase.java
@@ -19,11 +19,27 @@ import java.util.List;
 import org.json.JSONObject;
 
 import org.junit.After;
+import org.junit.BeforeClass;
 
 /**
  * @author Kenji Heigel
  */
 public abstract class BaseRelevantRuleTestCase {
+
+	@BeforeClass
+	public static void setUpClass() throws Exception {
+		jsonObject.put(
+			"build_profile", "DXP"
+		).put(
+			"git_repository_dir", "liferay-portal"
+		).put(
+			"job_name", "test-portal-acceptance-pullrequest("
+		).put(
+			"test_suite_name", "relevant"
+		).put(
+			"upstream_branch_name", "master"
+		);
+	}
 
 	@After
 	public void tearDown() {
@@ -117,6 +133,8 @@ public abstract class BaseRelevantRuleTestCase {
 
 		return relevantRuleEngine;
 	}
+
+	protected static JSONObject jsonObject = new JSONObject();
 
 	private File _baseDir;
 	private PortalAcceptancePullRequestJob _portalAcceptancePullRequestJob;
