+++ modules/test/jenkins-results-parser/src/test/java/com/liferay/jenkins/results/parser/test/suite/RelevantTestSuiteTest.java
@@ -23,10 +23,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import org.json.JSONObject;
-
 import org.junit.Assert;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 /**
@@ -34,25 +31,10 @@ import org.junit.Test;
  */
 public class RelevantTestSuiteTest extends BaseRelevantRuleTestCase {
 
-	@BeforeClass
-	public static void setUpClass() throws Exception {
-		_jsonObject.put(
-			"build_profile", "DXP"
-		).put(
-			"git_repository_dir", "liferay-portal"
-		).put(
-			"job_name", "test-portal-acceptance-pullrequest("
-		).put(
-			"test_suite_name", "relevant"
-		).put(
-			"upstream_branch_name", "master"
-		);
-	}
-
 	@Test
 	public void testJUnitTestSelectorMerge() throws IOException {
 		RelevantTestSuite relevantTestSuite = new RelevantTestSuite(
-			getPortalAcceptancePullRequestJob(_jsonObject));
+			getPortalAcceptancePullRequestJob(jsonObject));
 
 		relevantTestSuite.setModifiedFiles(
 			Arrays.asList(
@@ -100,7 +82,7 @@ public class RelevantTestSuiteTest extends BaseRelevantRuleTestCase {
 	@Test
 	public void testPlaywrightTestSelectorMerge() {
 		RelevantTestSuite relevantTestSuite = new RelevantTestSuite(
-			getPortalAcceptancePullRequestJob(_jsonObject));
+			getPortalAcceptancePullRequestJob(jsonObject));
 
 		relevantTestSuite.setModifiedFiles(
 			Arrays.asList(
@@ -144,7 +126,7 @@ public class RelevantTestSuiteTest extends BaseRelevantRuleTestCase {
 	@Test
 	public void testPoshiTestSelectorMerge() throws IOException {
 		RelevantTestSuite relevantTestSuite = new RelevantTestSuite(
-			getPortalAcceptancePullRequestJob(_jsonObject));
+			getPortalAcceptancePullRequestJob(jsonObject));
 
 		relevantTestSuite.setModifiedFiles(
 			Arrays.asList(
@@ -178,7 +160,5 @@ public class RelevantTestSuiteTest extends BaseRelevantRuleTestCase {
 			}
 		}
 	}
-
-	private static final JSONObject _jsonObject = new JSONObject();
 
 }
